Add login as user functionality

+++ app/controllers/sessions_controller.rb
@@ -11,7 +11,9 @@ class SessionsController < ApplicationController
       # using it to indicate the most recent login
       user.touch
 
-      session[:user_id] = user.id.to_s
+      set_user_session(user)
+      set_admin_session(user) if user.is_admin?
+
       redirect_to root_path, notice: 'Successfully logged in!'
     else
       flash.now.alert = 'Invalid credentials, please try again.'
@@ -20,8 +22,23 @@ class SessionsController < ApplicationController
   end
 
   def destroy
-    session.delete(:user_id)
+    if logged_in_as_other_user?
+      other_user = current_user
 
-    redirect_to login_path, notice: 'Logged out!'
+      set_user_session(current_admin_user)
+
+      redirect_to admins_path, notice: "No longer logged in as #{other_user.email}"
+    else
+      session.delete(:user_id)
+      session.delete(:admin_user_id) if session[:admin_user_id]
+
+      redirect_to login_path, notice: 'Logged out!'
+    end
+  end
+
+  private
+
+  def logged_in_as_other_user?
+    current_admin_user.present? && session[:user_id] != session[:admin_user_id]
   end
 end
